Succeed sidelink transmission for PSCCH

+++ src/lte/model/lte-ue-mac.h
@@ -472,7 +472,7 @@ private:
     SidelinkGrantV2V m_nextGrantV2V;
     SidelinkGrantV2V m_prevGrantV2V;
 
-    uint8_t m_reserveCount;
+    uint32_t m_reserveCount;
 
     std::list<SidelinkCommResourcePool::SidelinkTransmissionInfo> m_pscchTx; ///< List of PSCCH transmissions within the pool
     std::list<SidelinkCommResourcePool::SidelinkTransmissionInfo> m_psschTx; ///< List of PSSCH transmissions within the pool
